refactor: rebase js modules to folder lib
