+++ lib/errors/myError.js
