+++ lib/streams/streams.js
