+++ lib/validators/cmdCheck.js
