+++ lib/validators/fileCheck.js
