+++ lib/validators/fileValidator.js
