+++ package.json
@@ -11,7 +11,6 @@
     "author": "",
     "license": "ISC",
     "devDependencies": {
-        "eslint": "^8.1.0",
         "jest": "^27.3.1"
     }
 }
